+++ lib/Culqi/Client.php
@@ -3,38 +3,27 @@ namespace Culqi;
 
 use Culqi\Error as Errors;
 
-
-class Client
-{
-    /**
-    * La versión de API usada
-    */
-    const API_VERSION = "v1.2";
-
-    /**
-     * La URL Base por defecto
-     */
-    const BASE_URL = "https://integ-pago.culqi.com/api/v1";
-
-
+/**
+ * Class Client
+ *
+ * @package Culqi
+ */
+class Client {
     public function request($method, $url, $api_key, $data = NULL) {
         try {
-
             $headers= array("Authorization" => "Bearer ".$api_key, "Content-Type" => "application/json", "Accept" => "application/json");
-
             $options = array(
                 'timeout' => 120
             );
-
             if($method == "GET") {
                 $url_params = is_array($data) ? '?' . http_build_query($data) : '';
-                $response = \Requests::get(Culqi::$api_base . $url . $url_params, $headers, $options);
+                $response = \Requests::get(Culqi::BASE_URL. $url . $url_params, $headers, $options);
             } else if($method == "POST") {
-                $response = \Requests::post(Culqi::$api_base . $url, $headers, json_encode($data));
+                $response = \Requests::post(Culqi::BASE_URL . $url, $headers, json_encode($data), $options);
             } else if($method == "PATCH") {
-                $response = \Requests::patch(Culqi::$api_base . $url, $headers, json_encode($data), $options);
+                $response = \Requests::patch(Culqi::BASE_URL . $url, $headers, json_encode($data), $options);
             } else if($method == "DELETE") {
-                $response = \Requests::delete(Culqi::$api_base, $options);
+                $response = \Requests::delete(Culqi::BASE_URL, $options);
             }
         } catch (\Exception $e) {
             throw new Errors\UnableToConnect();
@@ -46,8 +35,6 @@ class Client
             return json_decode($response->body);
         }
         if ($response->status_code == 400) {
-            $code = 0;
-            $message = "";
             throw new Errors\UnhandledError($response->body, $response->status_code);
         }
         if ($response->status_code == 401) {