+++ lib/Culqi/Resource.php
@@ -1,8 +1,17 @@
 <?php
+
 namespace Culqi;
 
+/**
+ * Class Resource
+ *
+ * @package Culqi
+ */
 class Resource extends Client {
 
+    /**
+     * Constructor.
+     */
     public function __construct($culqi)
     {
         $this->culqi = $culqi;